requirements.txt
newsapi-python
python-dotenv
requests
streamlit
openai
groq